requirements.txt
langchain
langchain-community
pypdf
pinecone
sentence-transformers
torch
transformers
accelerate
sentencepiece
ollama
python-dotenv  